addressed pr review 2
