+++ src/components/BaseComponents/BaseScrollView.tsx
@@ -6,6 +6,7 @@ const styles = () =>
     scroll: {
       overflow: 'scroll',
       height: '100%',
+      width: '100%',
       paddingBottom: '15px',
     },
   });
